ai/project_root/requirements.txt
torch==2.4.1
torchvision==0.15.2
whisper
ffmpeg-python
librosa
pandas
numpy
PyYAML

